HDH-184-fixing-deep-linking-of-details
